* vulkan_compute: VK_EXT_inline_uniform_block is a required extension now (will be used soon for parameters)  * vulkan_program/vulkan_kernel: we don't actually need dynamic SSBOs, normal ones suffice -> remove dynamic offsets which aren't needed now  * universal_binary: previous change didn't work out as hoped (now leads to double-init/overwrite if not used properly) -> rather put the version and type into each backend struct (and make COMPUTE_TYPE and version uint64_t) + added common struct that also contains the version and type

+++ compute/compute_common.hpp
@@ -23,7 +23,7 @@
 #include <cstdint>
 
 //! used to differentiate between the different compute implementations
-enum class COMPUTE_TYPE : uint32_t {
+enum class COMPUTE_TYPE : uint64_t {
 	NONE	= 0u,
 	OPENCL	= 1u,
 	CUDA	= 2u,
